+++ packages/twenty-server/src/core/messaging/constants/messaging.constants.ts
