+++ packages/twenty-server/src/core/messaging/dtos/timeline-thread-participant.dto.ts
